backend/requirements.txt
google-cloud-recommender
google-cloud-monitoring
google-cloud-asset
google-cloud-resource-manager
google-cloud-compute
fastapi
uvicorn

